# LAB/LAB_04/server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Film Manager API
  description: |-
    REST APIs for the interaction with the Film Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:3001/
    description: "Local Development Server"
...
paths:
...
  /api/films/public/{filmId}/images/{imageId}:
    get:
      tags:
        - api/films/public/{filmId}/images/{imageId}
      security:
        - cookieAuth: []
      summary: Retrieve an image associated to a public film
      description: The image with ID `imageId`, associated to the film with ID `filmId`, is retrieved. In particular, by specifying the desired content type via the Accept header, the user can decide whether to retrieve the image data structure (json content type), which does not contain the image file, or the image file itself, in one of the supported image content types (image/png, image/jpg, image/jpeg and image/gif). Only the film owner or a film reviewer can perform this operation.
      operationId: getSingleImage
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
              examples:
                Image Created Successfully:
                  value:
                    id: 1
                    originalname: "poster"
                    filename: "image-123"
                    filmId: 3
                    self: "api/films/public/3/images/1"
                    delete: "api/films/public/3/images/1"
            image/png:
              schema:
                type: string
                format: binary
            image/jpg:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '403':
          description: Forbidden (user is not the owner or is not a reviewer of)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Access is forbidden"
                  value:
                    code: 403
                    message: "you do not have permission to access this resource."
        '404':
          description: Not Found (e.g., film or image with this id is not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Resource not found"
                  value:
                    code: 404
                    message: "the requested resource could not be found."
        '415':
          description: Unsupported Media Type (e.g., the request payload format is not supported)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unsupported media type"
                  value:
                    code: 415
                    message: "The media type of the request payload is not supported by the server."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
    delete:
      tags:
        - api/films/public/{filmId}/images/{imageId}
      security:
        - cookieAuth: []
      summary: Delete an image associated to a public film
      description:  |
        The image to be deleted is the one with ID `imageId`, associated to the film with ID `filmId`. Only the film owner can delete the image.
      operationId: deleteSingleImage
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
          content: {}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Authentication is required or failed"
                  value:
                    code: 401
                    message: "authentication is required. Please log in with valid credentials."
        '403':
          description: Forbidden (user is not the owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Access is forbidden"
                  value:
                    code: 403
                    message: "you do not have permission to access this resource."
        '404':
          description: Not Found (e.g., film or image with this id is not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Resource not found"
                  value:
                    code: 404
                    message: "the requested resource could not be found."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                default:
                  summary: "Unexpected error occurred"
                  value:
                    code: 500
                    message: "an unexpected error occurred on the server. Please try again later."
...
components:
  schemas:
...
    Image:
      description: "schema that represents an image of a public film"
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: "path to the JSON Schema to use for the validation of the JSON data structure describing the image"
        id:
          type: integer
          description: "The unique identifier for the image."
        originalname:
          type: string
          description: "The original name of the image file, excluding the file extension."
        filename:
          type: string
          description: "The generated filename for the uploaded image, typically including the file extension."
        filmId:
          type: integer
          description: "The unique identifier of the film associated with the image."
        # Links according to HATEOAS principle
        self:
          type: string
          format: uri
          description: "Link to the current image resource"
        delete:
          type: string
          format: uri
          description: "Link to delete the image resource"
      required:
        - id
        - name
        - type
...
    ErrorResponse:
      description: "Represents an error response returned by the API."
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: "path to the JSON Schema used for validating the error response structure"
        code:
          type: integer
          description: "error code"
        message:
          type: string
          description: "a brief description of the error"
      required:
        - code
        - message
  securitySchemes:
    cookieAuth:
      description: "JWT to identify the authenticated user."
      type: apiKey
      in: cookie
      name: connect.sid